# swagger z dodatkowym sectorem 2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GymApp",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Sectors": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sector"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sector"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sector"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sectors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sector"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Sector"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Sector"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Sector"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sector"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sector"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sectors/GetActiveSectors/{clientId}": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PassBoughtEvent"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PassBoughtEvent"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PassBoughtEvent"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Sectors/CheckIfActive": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sectorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sectors/{id}": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Sector"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sector"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sector"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sectors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sector"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Sector"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Sector"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Sectors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Entrance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "passId": {
            "type": "string",
            "format": "uuid"
          },
          "sectorId": {
            "type": "string",
            "format": "uuid"
          },
          "entranceEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntranceEvent"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EntranceEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "entranceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "PassBoughtEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "passId": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "refresh": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Sector": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sectorName": {
            "type": "string",
            "nullable": true
          },
          "gymId": {
            "type": "string",
            "format": "uuid"
          },
          "entrances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entrance"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "apiKey",
        "description": "Standard Authorization header using Bearer scheme (\"bearer {token}\")",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}